backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
kiteconnect
twilio
python-telegram-bot
emails
python-dotenv
websockets
redis>=5.0.0
sentry-sdk>=1.30.0
